feat: Process status reporting
